Fix minor GUI bugs and reorganize controllers into a package
