+++ src/main/java/duke/TaskType.java
@@ -1,0 +1,8 @@
+package duke;
+
+/**
+ * Enum for the three task types
+ */
+public enum TaskType {
+    TODO, EVENT, DEADLINE
+}
