+++ src/main/java/duke/controllers/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.controllers;
 
 import java.io.IOException;
 import java.util.Collections;
