+++ src/main/resources/view/DialogBox.fxml
@@ -6,14 +6,14 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<fx:root alignment="TOP_RIGHT" maxWidth="1.7976931348623157E308" minHeight="-Infinity" minWidth="-Infinity" prefHeight="108.0" prefWidth="392.0" style="-fx-background-color: #21130d; -fx-background-radius: 20px; -fx-opacity: 90%;" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1">
+<fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="-Infinity" prefHeight="108.0" prefWidth="392.0" style="-fx-background-color: #21130d; -fx-background-radius: 20px; -fx-opacity: 90%;" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1">
     <children>
-      <TextFlow maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308">
+      <TextFlow maxHeight="1.7976931348623157E308" maxWidth="-Infinity">
          <HBox.margin>
             <Insets left="5.0" right="5.0" />
          </HBox.margin>
          <children>
-              <Label fx:id="dialog" style="-fx-border-width: 3px; -fx-font-weight: bold; -fx-border-radius: 50px;" text="Label" textFill="WHITE" wrapText="true">
+              <Label fx:id="dialog" maxHeight="1.7976931348623157E308" maxWidth="-Infinity" style="-fx-border-width: 3px; -fx-font-weight: bold; -fx-border-radius: 50px; -fx-wrap-text: true;" text="Label" textFill="WHITE" wrapText="true">
                <font>
                   <Font name="Calibri" size="12.0" />
                </font>
